fix: Keep video tracks first
When MakeMKV produces a file, the video track is typically first
(track 0), followed by audio tracks, followed by subtitle tracks.

The `mkvmerge` remux step in the `entrypoint.sh` for the `dovi_tool`
Docker image reverses this order. Audio and subtitle tracks are first
and video tracks are last.

This commit fixes the order of files passed to `mkvmerge` so the video
track will remain track 0 in the remuxed file.